% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_idx.R
\name{aux_idx}
\alias{aux_idx}
\title{Auciliary indices for blocks}
\usage{
aux_idx(jj, idx_jk, no_eta)
}
\arguments{
\item{jj}{elements X for lpi}

\item{idx_jk}{indices for sparsity}

\item{no_eta}{number of lpi}
}
\value{
indices for blocks
}
\description{
Auxiliary indices
}
