% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_eff.R
\name{boost_eff}
\alias{boost_eff}
\title{Extract information from the boosting procedure}
\usage{
boost_eff(data_boost)
}
\arguments{
\item{data_boost}{the output of the boosting_step() function}
}
\value{
A data.frame with information on the importance of each effect
}
\description{
Extract information from the boosting procedure
}
