% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_step.R
\name{boosting_step}
\alias{boosting_step}
\title{Order effects by gradient boosting}
\usage{
boosting_step(
  data,
  data_v,
  effects,
  constraints,
  nstep,
  rate,
  diagonal,
  ncores,
  ncv,
  old_fit = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{training data}

\item{data_v}{validation data}

\item{constraints}{list of constraints}

\item{nstep}{number of boosting steps}

\item{rate}{learning rate}

\item{diagonal}{if TRUE only diagonal of MCD is modelled}

\item{ncores}{number of parallel thread to use}

\item{ncv}{number of fold for cross-validation}

\item{old_fit}{can be use to start boosting from an old boosting run}

\item{verbose}{print information as we boost?}

\item{effect}{list of effects}
}
\value{
A list containing information on the effects and their importance
}
\description{
Order effects by gradient boosting
}
