% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_GAMLSS.R
\name{build_GAMLSS}
\alias{build_GAMLSS}
\title{Function for building new GAMLSS families}
\usage{
build_GAMLSS(o)
}
\arguments{
\item{o}{a bundled objects, such as bundle_GPD}
}
\value{
a family that can be used with mgcv::gam for fitting
}
\description{
Function for building new GAMLSS families
}
