% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_GPD.R
\docType{data}
\name{bundle_GPD}
\alias{bundle_GPD}
\title{Ingredients for Generalized Pareto family}
\format{
An object of class \code{list} of length 10.
}
\usage{
bundle_GPD
}
\description{
Ingredients for Generalized Pareto family
}
\keyword{datasets}
