% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_gau_copula.R
\name{d_gau_copula}
\alias{d_gau_copula}
\title{Density of Gaussian copula}
\usage{
d_gau_copula(logu, R, log = TRUE)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
