% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss.gH_mcd.R
\name{gamlss.gH_mcd}
\alias{gamlss.gH_mcd}
\title{New gamlss.gH_mcd}
\usage{
gamlss.gH_mcd(
  X,
  jj,
  eta,
  y,
  w,
  z,
  t,
  Gm,
  l1,
  l1_l,
  l2,
  l2_v,
  l2_l,
  l2_v_l,
  idx_b,
  idx_aux,
  l3 = 0,
  i3 = 0,
  l4 = 0,
  i4 = 0,
  d1b = 0,
  d2b = 0,
  deriv = 0,
  fh = NULL,
  D = NULL
)
}
\arguments{
\item{X}{model matrix}

\item{jj}{indices X for lpi}

\item{eta}{lpi}

\item{y}{outcome matrix}

\item{w}{indices vector}

\item{z}{indices vector}

\item{t}{indices vector}

\item{Gm}{indices matrix}

\item{l1}{initialised score eta}

\item{l1_l}{initialised score eta (last block)}

\item{l2}{initialised hessian eta}

\item{l2_v}{initialised hessian eta (cumulative)}

\item{l2_l}{initialised hessian eta (last block)}

\item{l2_v_l}{initialised hessian eta (cumulative - last block)}

\item{idx_b}{block indices}

\item{idx_aux}{list sparsity indices}

\item{l3}{3rd der}

\item{i3}{3rd der indices}

\item{l4}{4th der}

\item{i4}{4th der indices}

\item{d1b}{1st der beta}

\item{d2b}{2nd der beta}

\item{deriv}{integer for optimisation routine}

\item{fh}{Hessian inverse}

\item{D}{matrix ...}
}
\value{
derivatives
}
\description{
Routine for computing the derivatives
}
