% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov_plot.R
\name{get_cov_plot}
\alias{get_cov_plot}
\title{Plot covariance matrices as a network}
\usage{
get_cov_plot(indexes, CORRS, DIAG, A)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
