% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effects.R
\name{get_effects}
\alias{get_effects}
\title{Creates list of effects for gradient boosting}
\usage{
get_effects(version, allNs, diagonal)
}
\arguments{
\item{version}{what type of model are we using "c", "c+w" etc}

\item{allNs}{vector of names of GSP groups}

\item{diagonal}{if TRUE the effects will be used to model only diagonal of MCD}
}
\value{
List of effects to be passed to boosting_step() function
}
\description{
Creates list of effects for gradient boosting
}
