% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norm_resid.R
\name{get_norm_resid}
\alias{get_norm_resid}
\title{Get normalized residuals from GAMLSS model}
\usage{
get_norm_resid(x, d)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
