% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resids_in_out.R
\name{get_resids_in_out}
\alias{get_resids_in_out}
\title{Get uniform residuals from GAMLSS models on test and training sets}
\usage{
get_resids_in_out(mod, final, logp)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
