% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_scores}
\alias{get_scores}
\title{Get performance scores}
\usage{
get_scores(x, d, nsim, ncores, A)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
