% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores_copula.R
\name{get_scores_copula}
\alias{get_scores_copula}
\title{Get performance scores}
\usage{
get_scores_copula(fit_lss, nsim, ncores, A, indep = FALSE)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
