% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores_copula_gpd.R
\name{get_scores_copula_gpd}
\alias{get_scores_copula_gpd}
\title{Get performance scores from shash+gpd+copula model}
\usage{
get_scores_copula_gpd(
  fit_shash,
  fit_gpd,
  sets,
  nsim,
  ncores,
  A,
  dyn_copula = NULL,
  indep = FALSE
)
}
\arguments{
\item{fit_shash}{the estimates provided by the marginal GAMLSS models}

\item{fit_gpd}{the estimates provided by the marginal GPD models}

\item{sets}{the splitting points for the test set}

\item{nsim}{the number of simulation use to compute (e.g.) the CRPS loss}

\item{ncores}{the number of cores}

\item{A}{a matrix used to specify the linear combination of regions we are looking}

\item{dyn_copula}{if TRUE the copula is assumed to be dynamic (experimental)}

\item{indep}{if TRUE then the regions are assumed to be independent under the model}
}
\value{
A list of performance scores.
}
\description{
Get performance scores from shash+gpd+copula model
}
