% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shash_norm_resid.R
\name{get_shash_norm_resid}
\alias{get_shash_norm_resid}
\title{Get normalized residuals from shash model}
\usage{
get_shash_norm_resid(x, d, ncores)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
