% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_scores.R
\name{get_validation_scores}
\alias{get_validation_scores}
\title{Get validation scores}
\usage{
get_validation_scores(nstop_cross)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
