% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_pdf.R
\name{gpd_pdf}
\alias{gpd_pdf}
\title{The pdf of the GPD distribution}
\usage{
gpd_pdf(y, eta, log)
}
\arguments{
\item{y}{the data on which the pdf is evaluated}

\item{eta}{matrix of linear predictors}

\item{log}{if TRUE the log-pdf is returned}
}
\value{
the pdf corresponding to each y
}
\description{
The pdf of the GPD distribution
}
