% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idxHess_no0.R
\name{idxHess_no0}
\alias{idxHess_no0}
\title{Auciliary indices for sparsity}
\usage{
idxHess_no0(no_eta, z, w, param)
}
\arguments{
\item{no_eta}{number of lpi}

\item{z}{vector z}

\item{w}{vector w}

\item{param}{mcd parametrisation}
}
\value{
indices for blocks
}
\description{
Auxiliary indices
}
