% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx0s.R
\name{index0s}
\alias{index0s}
\title{Index Zero components}
\usage{
index0s(no_eta, z, w, Gmat, t)
}
\arguments{
\item{no_eta}{The dimension of the outcome}

\item{z}{auxiliary index vector z}

\item{w}{auxiliary index vector w}

\item{Gmat}{auxiliary index matrix G (not used here)}

\item{t}{auxiliary index vector t (not used here)}
}
\value{
list of zero elements for each row
}
\description{
The G matrix
}
