% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_data_for_gams.R
\name{initialise_data_for_gams}
\alias{initialise_data_for_gams}
\title{Initialise data for fitting}
\usage{
initialise_data_for_gams(dat, d)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
