% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_to_resp.R
\name{lpi_to_resp}
\alias{lpi_to_resp}
\title{Get covariance matrix from vector}
\usage{
lpi_to_resp(lpi, d)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
