% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_MCD.R
\name{mvn_mcd}
\alias{mvn_mcd}
\title{Family definition (internal use)}
\usage{
mvn_mcd(d = 2, nb = 1)
}
\arguments{
\item{d}{dimension of the outcome}

\item{nb}{number of chunks}
}
\value{
family
}
\description{
Family function
}
