% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_gpd_fits.R
\name{postprocess_gpd_fits}
\alias{postprocess_gpd_fits}
\title{Post process marginal GPD fits}
\usage{
postprocess_gpd_fits(fit_s, fit_g, only_shash = FALSE)
}
\arguments{
\item{fit_s}{a list of fitted marginal shash models}

\item{fit_g}{a list of fitted marginl gam models}

\item{only_shash}{if TRUE the GPD distribution will not be used}
}
\value{
A list containing the fit obtained by a shash+gpd model
}
\description{
Post process marginal GPD fits
}
