% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash_qf.R
\name{shash_qf}
\alias{shash_qf}
\title{Quantile function of the shash distribution}
\usage{
shash_qf(p, mu, wt, scale, logp)
}
\arguments{
\item{p}{the probability level in (0, 1)}

\item{mu}{a matrix of shash parameters}

\item{wt}{a vector of weights}

\item{scale}{the scale parameter (ignored here)}

\item{logp}{if TRUE then p is log(p)}
}
\value{
a vector of quantiles
}
\description{
Quantile function of the shash distribution
}
