% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_step.R
\name{boosting_step}
\alias{boosting_step}
\title{Boosting ordering}
\usage{
boosting_step(
  data,
  data_v,
  effects,
  constraints,
  nstep,
  rate,
  diagonal,
  ncores,
  ncv = NULL,
  verbose = FALSE
)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
