% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_on_test_set.R
\name{fit_on_test_set}
\alias{fit_on_test_set}
\title{Fit MVN GAM on test set}
\usage{
fit_on_test_set(data_boost, nstop, data, start, ncv, ncores)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
