% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effects.R
\name{get_effects}
\alias{get_effects}
\title{Get effects and constraints}
\usage{
get_effects(version, allNs, diagonal)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
