% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_fits.R
\name{get_final_fits}
\alias{get_final_fits}
\title{Fit all MVN GAM on all data}
\usage{
get_final_fits(nstop_all, data, ncores, files)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
