% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_scores}
\alias{get_scores}
\title{Get performance scores for MVN GAMs on test set}
\usage{
get_scores(x, d, nsim, ncores, A)
}
\arguments{
\item{x}{a matrix of predictions from the MVN GAM}

\item{d}{the dimension of the response vector}

\item{nsim}{the number of simulation use to compute (e.g.) the CRPS loss}

\item{ncores}{the number of cores}

\item{A}{a matrix used to specify the linear combination of regions we are looking at}
}
\value{
A list of performance scores.
}
\description{
Get performance scores for MVN GAMs on test set
}
