% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores_copula_gpd.R
\name{get_scores_copula_gpd}
\alias{get_scores_copula_gpd}
\title{Get performance scores from shash+gpd+copula model}
\usage{
get_scores_copula_gpd(
  fit_shash,
  fit_gpd,
  sets,
  nsim,
  ncores,
  A,
  dyn_copula = NULL,
  indep = FALSE
)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
