% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_scores.R
\name{get_validation_scores}
\alias{get_validation_scores}
\title{Get validation scores to choose number of effects}
\usage{
get_validation_scores(nstop_cross)
}
\arguments{
\item{nstop_cross}{The output of the validation_step() function}
}
\value{
A vector of out-of-sample log-likelihood values
}
\description{
Get validation scores to choose number of effects
}
