% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_data_for_gams.R
\name{initialise_data_for_gams}
\alias{initialise_data_for_gams}
\title{Initialise UK data for fitting}
\usage{
initialise_data_for_gams(dat, d)
}
\arguments{
\item{dat}{the UK GSP group data}

\item{d}{the dimension (14 in this case)}
}
\value{
the data in the right format
}
\description{
Initialise UK data for fitting
}
