% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_to_resp.R
\name{lpi_to_resp}
\alias{lpi_to_resp}
\title{Convert linear predictors (on MCD scale) to covariance matrix}
\usage{
lpi_to_resp(lpi, d)
}
\arguments{
\item{lpi}{a matrix of linear predictors}

\item{d}{the dimension of the response vector}
}
\value{
a matrix with means and covariances on each row
}
\description{
Convert linear predictors (on MCD scale) to covariance matrix
}
