% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_gpd_fits.R
\name{postprocess_gpd_fits}
\alias{postprocess_gpd_fits}
\title{Post process marginal GPD fits}
\usage{
postprocess_gpd_fits(fit_s, fit_g, only_shash = FALSE)
}
\arguments{
\item{X}{XXX}
}
\value{
XXX
}
\description{
XXX
}
