% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_step.R
\name{validation_step}
\alias{validation_step}
\title{Validation for optimal number of effects}
\usage{
validation_step(data_boost, dat, start, nstop, ncv, ncores)
}
\arguments{
\item{theta}{B}
}
\value{
C
}
\description{
A
}
\details{
This function is meant for internal use only.
}
